Update SLA schema tests
