+++ tests/schemas/test_sla_schema_cases.py
@@ -1,0 +1,24 @@
+from typing import Literal
+
+from pytest_cases import case, parametrize
+
+from tests.utils import random_lower_string
+
+
+class CaseAttr:
+    @case(tags=["base", "base_public", "update"])
+    def case_none(self) -> tuple[None, None]:
+        return None, None
+
+    @case(tags=["update"])
+    def case_attr(self) -> tuple[Literal["doc_uuid"], None]:
+        return "doc_uuid", None
+
+    @case(tags=["update"])
+    @parametrize(attr=["start_date", "end_date"])
+    def case_nullable_dates(self, attr: str) -> tuple[str, None]:
+        return attr, None
+
+    @case(tags=["base", "base_public"])
+    def case_desc(self) -> tuple[Literal["description"], str]:
+        return "description", random_lower_string()
